fix: carry detached artifact verification
